refactor(serverless-plugin): migrate from https.get to fetch
